chore: deploy config

+++ Full-Stack/Projetos Estudos/Rocket Log/src/tests/users-controller.test.ts
@@ -1,40 +1,48 @@
 import request from "supertest";
-import { app } from "../app";
-import { after } from "node:test";
 import { prisma } from "@/database/prisma";
+
+import { app } from "@/app";
 
 describe("UsersController", () => {
   let user_id: string;
 
   afterAll(async () => {
-    await prisma.user.delete({
-      where: {
-        id: user_id,
-      },
-    });
+    await prisma.user.delete({ where: { id: user_id } });
   });
 
-  it("should create a new user", async () => {
+  it("should create a new user successfully", async () => {
     const response = await request(app).post("/users").send({
-      name: "Lucas test",
-      email: "testuser2@example.com",
-      password: "123456789",
+      name: "Test User",
+      email: "testuser@example.com",
+      password: "password123",
     });
+
     expect(response.status).toBe(201);
     expect(response.body).toHaveProperty("id");
-    expect(response.body).toHaveProperty("name");
-    expect(response.body).toHaveProperty("email");
+    expect(response.body.name).toBe("Test User");
 
     user_id = response.body.id;
   });
 
-  it("should not create a user with an existing email", async () => {
+  it("should throw an error if user with same email already exists", async () => {
     const response = await request(app).post("/users").send({
-      name: "Lucas test",
-      email: "testuser2@example.com",
-      password: "123456789",
+      name: "Duplicate User",
+      email: "testuser@example.com",
+      password: "password123",
     });
+
     expect(response.status).toBe(400);
-    expect(response.body).toHaveProperty("error");
+    expect(response.body.message).toBe("User with same email already exists");
+  });
+
+  it("should throw a validation error if email is invalid", async () => {
+    const response = await request(app).post("/users").send({
+      name: "Test User",
+      email: "invalid-email",
+      password: "password123",
+    });
+
+    expect(response.status).toBe(400);
+    expect(response.body.message).toBe("validation error");
   });
 });
